requirements.txt
timm==0.6.11
numpy==1.21.6
torch=1.12.1+cu113
torchvison==0.13.1+cu113
opencv-python==4.6.0
matplotlib==3.2.2
Pillow==7.1.2
json==2.0.9